+++ Back-End (Spring)/demo/src/main/java/com/AgendaServico/demo/Repository/ClienteRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository // Anotacao informa para o Spring que essa classe acessa o banco
-public interface ClienteRepository extends JpaRepository<Cliente, Long>{
+public interface ClienteRepository extends JpaRepository<Cliente, Integer>{
 
     // extends JpaRepository<...> --> Importa os métodos prontos do CRUD
-    // <Cliente, Long> --> entidade que sera gerenciada e tipo de chave primaria (id)
+    // <Cliente, Integer> --> entidade que sera gerenciada e tipo de chave primaria (id)
 
 }
