+++ Back-End (Spring)/demo/src/main/java/com/AgendaServico/demo/Repository/ServicoRepository.java
@@ -6,5 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ServicoRepository extends JpaRepository<Servico, Long> {
+public interface ServicoRepository extends JpaRepository<Servico, Integer> {
+
+    // extends JpaRepository<...> --> Importa os métodos prontos do CRUD
+    // <Servico, Integer> --> entidade que sera gerenciada e tipo de chave primaria (id)
+
 }
